添加淡入淡出效果

+++ src/utils/tweenMax/fragment.fs
@@ -2,7 +2,7 @@ varying vec2 vUv;
 uniform sampler2D currentImage;
 uniform sampler2D nextImage;
 uniform float dispFactor;
-uniform float switchType;
+uniform int switchType;
 uniform bool reverse;
 
 //branchDefines
